Refactor parsing and formatting functions to "parsing" module
